requirements.txt
requests
FastAPI
pyYAML
joblib
types-PyYAML
pandas
ensure
defusedxml
uvicorn
python-box
setuptools

-e .